Stable, as I think

+++ TunerBass/Graph.h
@@ -30,5 +30,5 @@ public:
 	void clear();				//paint it black
 	void draw_axis();			//draw coordinate plane					
 	void draw_new_point(float x, float y);		//add new point to the graph
-	void draw_sample(int* sample, int offset);	//output sample to the graph
+	void draw_sample(float* sample, int offset);	//output sample to the graph
 };
